+++ server/app.py
@@ -18,17 +18,21 @@ from models.alert import PlantAlert
 from models.todo import Todo
 from models.mix import Mix, Soil, SoilPart
 
-from db import init_db
-from db import Session
+from db import init_db, Session
 
 from logger import setup_logger
 import logging
+
+from discover import discover_systems
 
 # Create a logger for this specific module
 logger = setup_logger(__name__, logging.DEBUG)
 
 # Initialize DB connection
 init_db()
+
+#
+discover_systems() # Maybe put this into the installable?
 
 # Create Flask app
 app = Flask(__name__)
@@ -64,58 +68,6 @@ app.register_blueprint(soils_bp)
 
 CORS(app)
 
-ENVIRONMENT = os.getenv('ENVIRONMENT', 'docker')
-USE_LOCAL_CAMERA = os.getenv('USE_LOCAL_CAMERA', 'false').lower() == 'true'
-
-def discover_systems():
-    session = Session()
-    
-    if USE_LOCAL_CAMERA:
-        local_camera = session.query(System).filter_by(name='Local Camera').first()
-        if not local_camera:
-            local_camera = Camera(name='Local Camera', url='local', container_id='local')
-            session.add(local_camera)
-    # We'll get there...
-    # if ENVIRONMENT == 'kubernetes':
-    #     from kubernetes import client, config
-    #     config.load_incluster_config()
-    #     v1 = client.CoreV1Api()
-    #     pods = v1.list_pod_for_all_namespaces(label_selector="app=pi-camera").items
-        
-    #     for pod in pods:
-    #         pi_id = pod.spec.node_name
-    #         pod_ip = pod.status.pod_ip
-    #         url = f"http://{pod_ip}:5000"
-            
-    #         camera = session.query(Camera).filter_by(name=f"Pi Camera {pi_id}").first()
-    #         if not camera:
-    #             camera = Camera(name=f"Pi Camera {pi_id}", url=url, container_id=pod.metadata.uid)
-    #             session.add(camera)
-    #         else:
-    #             camera.url = url
-    #             camera.container_id = pod.metadata.uid
-    
-    elif ENVIRONMENT == 'docker':
-        import docker
-        client = docker.from_env()
-        
-        for container in client.containers.list():
-            if container.name.startswith('pi-camera-'):
-                pi_id = container.name.split('-')[-1]
-                ip = container.attrs['NetworkSettings']['Networks']['multi-camera_default']['IPAddress']
-                url = f"http://{ip}:5000"
-                
-                camera = session.query(Camera).filter_by(container_id=container.id).first()
-                if not camera:
-                    camera = Camera(name=f"Pi Camera {pi_id}", url=url, container_id=container.id)
-                    session.add(camera)
-                else:
-                    camera.url = url
-    
-    session.commit()
-    cameras = session.query(Camera).all()
-    session.close()
-    return cameras
 
 
 @app.route("/meta/", methods=["GET"])
